[health-check-endpoint]
Refactor into main/lib.rs and added skeleton for testing
