+++ tests/health_check.rs
@@ -1,0 +1,7 @@
+//! tests/health_check.rs
+
+use zero2prod::main;
+#[test]
+fn dummy_test() {
+    main();
+}
